fix mac warnings
git-svn-id: https://delta.affinix.com/svn/trunk/psimedia@521 ac7a513c-5e3a-0410-af42-968c3ac6e7dc

+++ gstprovider/gstelements/liveadder/liveadder.c
@@ -960,7 +960,7 @@ gst_live_live_adder_chain (GstPad *pad, GstBuffer *buffer)
         " (expected %" GST_TIME_FORMAT ", got %" GST_TIME_FORMAT" drift:%ldms)",
         GST_TIME_ARGS (padprivate->expected_timestamp),
 		    GST_TIME_ARGS (GST_BUFFER_TIMESTAMP (buffer)),
-		    drift / GST_MSECOND);
+		    (long int)(drift / GST_MSECOND));
 
     /* We accept drifts of 10ms*/
     if (ABS(drift) < (10 * GST_MSECOND)) {
